basix3

+++ src/bfqs.h
@@ -35,8 +35,8 @@ public:
     uint8_t *last_t;
     uint8_t last_leaf_child;
 
-    bfqs(uint16_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
-            uint16_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
+    bfqs(uint32_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
+            uint32_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
             const char *fname = NULL, uint8_t *block = NULL) :
         bpt_trie_handler<bfqs>(leaf_block_sz, parent_block_sz, cache_sz, fname, block) {
     }
