+++ src/dfos.h
@@ -50,8 +50,8 @@ public:
 #endif
     const static uint8_t need_counts[10];
 
-    dfos(uint16_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
-            uint16_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
+    dfos(uint32_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
+            uint32_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
             const char *fname = NULL, uint8_t *block = NULL) :
         bpt_trie_handler<dfos>(leaf_block_sz, parent_block_sz, cache_sz, fname, block) {
     }
