+++ src/dfqx.h
@@ -37,8 +37,8 @@ public:
 public:
     uint8_t pos, key_at_pos;
 
-    dfqx(uint16_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
-            uint16_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
+    dfqx(uint32_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
+            uint32_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
             const char *fname = NULL, uint8_t *block = NULL) :
         bpt_trie_handler<dfqx>(leaf_block_sz, parent_block_sz, cache_sz, fname, block) {
     }
