+++ src/dft.h
@@ -20,8 +20,8 @@ public:
     int16_t last_sibling_pos;
     int16_t pos, key_at_pos;
 
-    dft(uint16_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
-            uint16_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
+    dft(uint32_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
+            uint32_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
             const char *fname = NULL, uint8_t *block = NULL) :
         bpt_trie_handler<dft>(leaf_block_sz, parent_block_sz, cache_sz, fname, block) {
     }
