+++ src/octp.h
@@ -47,8 +47,8 @@ public:
     uint8_t key_char_pos;
     uint8_t bit_map_len;
 
-    octp(uint16_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
-            uint16_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
+    octp(uint32_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
+            uint32_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
             const char *fname = NULL, uint8_t *block = NULL) :
         bpt_trie_handler<octp>(leaf_block_sz, parent_block_sz, cache_sz, fname, block) {
     }
