+++ src/rb_tree.h
@@ -64,8 +64,8 @@ public:
     int16_t pos;
     uint8_t last_direction;
 
-    rb_tree(uint16_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
-            uint16_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
+    rb_tree(uint32_t leaf_block_sz = DEFAULT_LEAF_BLOCK_SIZE,
+            uint32_t parent_block_sz = DEFAULT_PARENT_BLOCK_SIZE, int cache_sz = 0,
             const char *fname = NULL, uint8_t *block = NULL) :
        bplus_tree_handler<rb_tree>(leaf_block_sz, parent_block_sz, cache_sz, fname, block) {
         GenTree::generateLists();
